+++ bindgen/src/templates/generic/ffi/ErrorTemplate.kt
@@ -64,3 +64,8 @@ object {{ e|ffi_converter_name }} : FfiConverterRustBuffer<{{ type_name }}> {
     }
 
 }
+
+{#- The `actual` side of the shims `generic/common/ErrorTemplate.kt` declared for this
+    error enum's methods and uniffi trait exports; see `self_shim_expect` in `macros.kt`. -#}
+{%- let uniffi_trait_methods = e.uniffi_trait_methods() %}
+{%- call kt::self_shims_actual(e.methods(), uniffi_trait_methods, type_name, false) %}{% endcall %}